use a TLSHandshake timeout in go 1.3+, where it's available
