+++ http.go
@@ -11,7 +11,6 @@ import (
 	"net/http"
 	"strings"
 	"sync"
-	"time"
 )
 
 var insecureClient = (*http.Client)(nil)
@@ -69,23 +68,6 @@ func GetNonValidatingHTTPClient() *http.Client {
 			InsecureSkipVerify: true,
 		}),
 	}
-}
-
-// NewHttpTLSTransport returns a new http.Transport constructed with the TLS config
-// and the necessary parameters for Juju.
-func NewHttpTLSTransport(tlsConfig *tls.Config) *http.Transport {
-	// See https://code.google.com/p/go/issues/detail?id=4677
-	// We need to force the connection to close each time so that we don't
-	// hit the above Go bug.
-	transport := &http.Transport{
-		Proxy:               http.ProxyFromEnvironment,
-		TLSClientConfig:     tlsConfig,
-		DisableKeepAlives:   true,
-		Dial:                dial,
-		TLSHandshakeTimeout: 10 * time.Second,
-	}
-	registerFileProtocol(transport)
-	return transport
 }
 
 // BasicAuthHeader creates a header that contains just the "Authorization"
